LPS-190476 move FeatureFlag[Type, Constants] to portal-kernel
