+++ portal-kernel/src/com/liferay/portal/kernel/feature/flag/constants/FeatureFlagConstants.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.feature.flag.web.internal.constants;
+package com.liferay.portal.kernel.feature.flag.constants;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
