+++ portal-kernel/src/com/liferay/portal/kernel/feature/flag/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
