+++ portal-kernel/src/com/liferay/portal/kernel/feature/flag/packageinfo
@@ -1,1 +1,1 @@
-version 1.0.0
+version 2.0.0
